refactor: calamity document mapping 정의
