+++ src/main/java/focandlol/calamity/domain/CalamityDocument.java
@@ -2,7 +2,6 @@ package focandlol.calamity.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import focandlol.calamity.dto.CalamityMessageDto;
-import focandlol.calamity.dto.Region;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Set;
@@ -12,8 +11,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
+import org.springframework.data.elasticsearch.annotations.Field;
+import org.springframework.data.elasticsearch.annotations.FieldType;
+import org.springframework.data.elasticsearch.annotations.InnerField;
+import org.springframework.data.elasticsearch.annotations.MultiField;
+import org.springframework.data.elasticsearch.annotations.Setting;
 
 @Document(indexName = "calamity-read")
+@Setting(settingPath = "/elasticsearch/calamity-settings.json")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
@@ -22,22 +27,36 @@ import org.springframework.data.elasticsearch.annotations.Document;
 public class CalamityDocument {
 
   @Id
+  @Field(type = FieldType.Keyword)
   private String id;
 
+  @Field(type = FieldType.Text, analyzer = "custom_nori")
   private String message;
 
+  @Field(type = FieldType.Text, analyzer = "custom_nori")
   private String region;
 
+  @Field(type = FieldType.Keyword)
+  private Set<String> regionsSet;
+
+  @MultiField(mainField = @Field(type = FieldType.Text, analyzer = "custom_nori"),
+      otherFields = @InnerField(suffix = "keyword", type = FieldType.Keyword))
   private Set<String> sido;
 
+  @Field(type = FieldType.Nested)
   private Set<Region> regions;
 
+  @MultiField(mainField = @Field(type = FieldType.Text, analyzer = "custom_nori"),
+      otherFields = @InnerField(suffix = "keyword", type = FieldType.Keyword))
   private String category;
 
+  @Field(type = FieldType.Date)
   private String createdAt;
 
+  @Field(type = FieldType.Date)
   private String registeredDate;
 
+  @Field(type = FieldType.Date)
   private String modifiedDate;
 
   public static CalamityDocument from(CalamityMessageDto dto) {
@@ -55,7 +74,9 @@ public class CalamityDocument {
   }
 
   private static String formatIso8601(String raw) {
-    if (raw == null || raw.isBlank()) return null;
+    if (raw == null || raw.isBlank()) {
+      return null;
+    }
     DateTimeFormatter inputFormatter;
     if (raw.length() == 19) { // 2025/01/06 18:41:19
       inputFormatter = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
