+++ src/main/java/focandlol/calamity/service/CalamityService.java
@@ -3,15 +3,12 @@ package focandlol.calamity.service;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import focandlol.calamity.dto.CalamityMessageDto;
-import focandlol.calamity.dto.Region;
+import focandlol.calamity.domain.Region;
 import focandlol.calamity.dto.RegionData;
 import focandlol.calamity.repository.CalamityRepository;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
